+++ modules/store/sources-jvm/RaptorLogStore.kt
@@ -1,0 +1,9 @@
+package io.fluidsonic.raptor.store
+
+
+public interface RaptorLogStore<in Value : Any> {
+
+	public suspend fun append(value: Value)
+
+	public companion object
+}